:truck: api: Migrate session-for-supervising to session-management
