+++ api/src/certification/session-management/infrastructure/repositories/index.js
@@ -9,6 +9,7 @@ import * as certificationOfficerRepository from './certification-officer-reposit
 import * as finalizedSessionRepository from './finalized-session-repository.js';
 import * as jurySessionRepository from './jury-session-repository.js';
 import * as sessionForInvigilatorKitRepository from './session-for-invigilator-kit-repository.js';
+import * as sessionForSupervisingRepository from './session-for-supervising-repository.js';
 import * as sessionRepository from './session-repository.js';
 import * as supervisorAccessRepository from './supervisor-access-repository.js';
 
@@ -24,6 +25,7 @@ import * as supervisorAccessRepository from './supervisor-access-repository.js';
  * @typedef {sessionForInvigilatorKitRepository} SessionForInvigilatorKitRepository
  * @typedef {issueReportCategoryRepository} IssueReportCategoryRepository
  * @typedef {certificationIssueReportRepository} CertificationIssueReportRepository
+ * @typedef {sessionForSupervisingRepository} SessionForSupervisingRepository
  * @typedef {sessionRepository} SessionRepository
  * @typedef {supervisorAccessRepository} SupervisorAccessRepository
  * @typedef {certificationReportRepository} CertificationReportRepository
@@ -39,6 +41,7 @@ const repositoriesWithoutInjectedDependencies = {
   issueReportCategoryRepository,
   certificationIssueReportRepository,
   sessionRepository,
+  sessionForSupervisingRepository,
   supervisorAccessRepository,
   certificationReportRepository,
 };
