+++ api/tests/certification/session-management/unit/domain/usecases/get-session-for-supervising_test.js
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 
-import { constants } from '../../../../lib/domain/constants.js';
-import { getSessionForSupervising } from '../../../../lib/domain/usecases/get-session-for-supervising.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { constants } from '../../../../../../lib/domain/constants.js';
+import { getSessionForSupervising } from '../../../../../../src/certification/session-management/domain/usecases/get-session-for-supervising.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 const START_DATETIME_STUB = new Date('2022-10-01T13:00:00Z');
 const COMPLEMENTARY_EXTRATIME_STUB = 45;
