+++ api/tests/certification/session-management/unit/domain/usecases/supervise-session_test.js
@@ -1,6 +1,6 @@
-import { InvalidSessionSupervisingLoginError, SessionNotAccessible } from '../../../../lib/domain/errors.js';
-import { superviseSession } from '../../../../lib/domain/usecases/supervise-session.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { InvalidSessionSupervisingLoginError, SessionNotAccessible } from '../../../../../../lib/domain/errors.js';
+import { superviseSession } from '../../../../../../src/certification/session-management/domain/usecases/supervise-session.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | supervise-session', function () {
   let sessionRepository;
