+++ api/tests/unit/application/session/index_test.js
@@ -7,11 +7,9 @@ import streamToPromise from 'stream-to-promise';
 
 import { NotFoundError } from '../../../../lib/application/http-errors.js';
 import { authorization } from '../../../../lib/application/preHandlers/authorization.js';
-import { assessmentSupervisorAuthorization as sessionSupervisorAuthorization } from '../../../../lib/application/preHandlers/session-supervisor-authorization.js';
 import { finalizedSessionController } from '../../../../lib/application/sessions/finalized-session-controller.js';
 import * as moduleUnderTest from '../../../../lib/application/sessions/index.js';
 import { sessionController } from '../../../../lib/application/sessions/session-controller.js';
-import { sessionForSupervisingController } from '../../../../lib/application/sessions/session-for-supervising-controller.js';
 import { sessionWithCleaCertifiedCandidateController } from '../../../../lib/application/sessions/session-with-clea-certified-candidate-controller.js';
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
@@ -148,40 +146,6 @@ describe('Unit | Application | Sessions | Routes', function () {
 
       // then
       expect(response.statusCode).to.equal(404);
-    });
-  });
-
-  describe('GET /api/sessions/{id}/supervising', function () {
-    it('should return 200 if the user is a supervisor of the session', async function () {
-      //given
-      sinon.stub(sessionSupervisorAuthorization, 'verifyBySessionId').callsFake((request, h) => h.response(true));
-      sinon.stub(sessionForSupervisingController, 'get').returns('ok');
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/sessions/3/supervising');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return 401 if the user is not a supervisor of the session', async function () {
-      //given
-      sinon
-        .stub(sessionSupervisorAuthorization, 'verifyBySessionId')
-        .callsFake((request, h) => h.response().code(401).takeover());
-      sinon.stub(sessionForSupervisingController, 'get').returns('ok');
-
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/sessions/3/supervising');
-
-      // then
-      expect(response.statusCode).to.equal(401);
     });
   });
